+++ lang.txt
@@ -1,2 +1,1 @@
-go
 debug
