+++ src/core/arithmetic.go
