+++ src/core/pop.go
